This commit was manufactured by cvs2svn to create tag 'VRJ_RELENG_2_0_0_RELEASE'.
git-svn-id: https://vrjuggler.googlecode.com/svn/tags/VRJ_RELENG_2_0_0_RELEASE/juggler@17859 08b38cba-cd3b-11de-854e-f91c5b6e4272

+++ modules/vrjuggler/test/Draw/OGL/cubes/cubesApp.h
@@ -283,7 +283,7 @@ public:
    {
       // We can't start until there are users registered wth the system
       // We rely upon users to keep track of the multi-user data structure
-      int num_users = vrj::Kernel::instance()->getUsers().size();
+      size_t num_users = vrj::Kernel::instance()->getUsers().size();
       return (num_users > 0);
    }
